Move the compiler_context and parser_context structs out of the headers
